Add OpenAPI parameters
