+++ samples/CarterSample/Features/Actors/OpenApi/GetActorById.cs
@@ -1,6 +1,6 @@
 namespace CarterSample.Features.Actors
 {
-    using Carter;
+    using Carter.OpenApi;
 
     public class GetActorById : RouteMetaData
     {
