+++ samples/CarterSample/Features/Actors/OpenApi/GetActors.cs
@@ -1,7 +1,7 @@
 namespace CarterSample.Features.Actors
 {
     using System.Collections.Generic;
-    using Carter;
+    using Carter.OpenApi;
 
     public class GetActors : RouteMetaData
     {
@@ -22,5 +22,11 @@ namespace CarterSample.Features.Actors
         public override string OperationId { get; } = "Actors_GetActors";
 
         public override string SecuritySchema { get; set; } = "ApiKey";
+
+        public override QueryStringParameter[] QueryStringParameter { get; } =
+        {
+            new QueryStringParameter { Name = "Offset", Description = "How man items to offset from the start of the list", Type = typeof(int) },
+            new QueryStringParameter { Name = "Size", Description = "How many items to return at one time. MAX:20", Type = typeof(int) }
+        };
     }
 }
