+++ src/CarterModule.cs
@@ -3,6 +3,7 @@ namespace Carter
     using System;
     using System.Collections.Generic;
     using System.Threading.Tasks;
+    using Carter.OpenApi;
     using Microsoft.AspNetCore.Http;
     using Microsoft.AspNetCore.Routing;
 
@@ -76,7 +77,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Get<T>(string path, Func<HttpRequest, HttpResponse, RouteData, Task> handler) where T : RouteMetaData
         {
             Task RequestDelegate(HttpContext httpContext) => handler(httpContext.Request, httpContext.Response, httpContext.GetRouteData());
@@ -88,7 +89,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Get<T>(string path, RequestDelegate handler) where T : RouteMetaData
         {
             path = this.RemoveStartingSlash(path);
@@ -127,7 +128,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Post<T>(string path, Func<HttpRequest, HttpResponse, RouteData, Task> handler) where T : RouteMetaData
         {
             Task RequestDelegate(HttpContext httpContext) => handler(httpContext.Request, httpContext.Response, httpContext.GetRouteData());
@@ -139,7 +140,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Post<T>(string path, RequestDelegate handler) where T : RouteMetaData
         {
             path = this.RemoveStartingSlash(path);
@@ -177,7 +178,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Delete<T>(string path, Func<HttpRequest, HttpResponse, RouteData, Task> handler) where T : RouteMetaData
         {
             Task RequestDelegate(HttpContext httpContext) => handler(httpContext.Request, httpContext.Response, httpContext.GetRouteData());
@@ -189,7 +190,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Delete<T>(string path, RequestDelegate handler) where T : RouteMetaData
         {
             path = this.RemoveStartingSlash(path);
@@ -227,7 +228,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Put<T>(string path, Func<HttpRequest, HttpResponse, RouteData, Task> handler) where T : RouteMetaData
         {
             Task RequestDelegate(HttpContext httpContext) => handler(httpContext.Request, httpContext.Response, httpContext.GetRouteData());
@@ -239,7 +240,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Put<T>(string path, RequestDelegate handler) where T : RouteMetaData
         {
             path = this.RemoveStartingSlash(path);
@@ -277,7 +278,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Head<T>(string path, Func<HttpRequest, HttpResponse, RouteData, Task> handler) where T : RouteMetaData
         {
             Task RequestDelegate(HttpContext httpContext) => handler(httpContext.Request, httpContext.Response, httpContext.GetRouteData());
@@ -289,7 +290,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Head<T>(string path, RequestDelegate handler) where T : RouteMetaData
         {
             path = this.RemoveStartingSlash(path);
@@ -328,7 +329,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Patch<T>(string path, Func<HttpRequest, HttpResponse, RouteData, Task> handler) where T : RouteMetaData
         {
             Task RequestDelegate(HttpContext httpContext) => handler(httpContext.Request, httpContext.Response, httpContext.GetRouteData());
@@ -340,7 +341,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Patch<T>(string path, RequestDelegate handler) where T : RouteMetaData
         {
             path = this.RemoveStartingSlash(path);
@@ -378,7 +379,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Options<T>(string path, Func<HttpRequest, HttpResponse, RouteData, Task> handler) where T : RouteMetaData
         {
             Task RequestDelegate(HttpContext httpContext) => handler(httpContext.Request, httpContext.Response, httpContext.GetRouteData());
@@ -390,7 +391,7 @@ namespace Carter
         /// </summary>
         /// <param name="path">The path for your route</param>
         /// <param name="handler">The handler that is invoked when the route is hit</param>
-        /// <typeparam name="T">The <see cref="Carter.RouteMetaData"/> implementation for OpenApi use</typeparam>
+        /// <typeparam name="T">The <see cref="OpenApi.RouteMetaData"/> implementation for OpenApi use</typeparam>
         protected void Options<T>(string path, RequestDelegate handler) where T : RouteMetaData
         {
             path = this.RemoveStartingSlash(path);
