--- src/RouteMetaDataResponse.cs
@@ -1,13 +1,0 @@
-namespace Carter
-{
-    using System;
-
-    public class RouteMetaDataResponse
-    {
-        public int Code { get; set; }
-
-        public string Description { get; set; }
-
-        public Type Response { get; set; }
-    }
-}
